crawler/requirements.txt
scrapy==2.14.1
beautifulsoup4==4.14.3
python-dateutil==2.9.0.post0
Twisted==25.5.0
itemadapter==0.13.1
